The game board file
Created the game board that contains the instance of the board and
maintains the rules (making legal queen and arrow moves).

Contains a helper accessor function to return the value of a desired
square (the value being whatever occupies the square, or 0 if empty)

All functions expect values starting at a base-1 index and then converts
the input to base-0.  If there's an output, it returns the value in
base-1.

movePiece and shootArrow both accept upper and lowercase letters for
their alphanumeric parameterized functions (ex, shootArrow("a4", "C6")).

Any illegal moves simply return an empty string and the board remains
unchanged.  (with exception to firing an arrow illegally; the queen will
remain where she last moved.  The program will have to force the player
to make a legal move before continuing on to the next player's turn)