Objectified ALL THE THINGS
Rewrote the gameboard to an object oriented design.  Queens, arrows and
blank spaces are now all objects and the board is now a 2-d array of
these objects.  This was done to make keeping track of queen locations
and localized "empty square" scores easier.

All pieces descend from Gamepiece, whose constructor stores x and y
coordinates.

White and BlackQueens and Arrows are similar in that they store their
x/y coordinates and a piece value (1 for white, 2 for black, -1 for
arrows), but Queens have a move() method that changes their internal
coordinate values (used for keeping track of their locations easier
instead of scanning the whole board looking for queens)

Blank pieces are simply empty squares, but have in them the internal
"Queen move" and "King move" scores for white and black.  Blank pieces
also contain functions to return a local heuristic evaluation for that
square, to be used by a main heuristic function that sums and weights
the values over all blank squares.

Blanks TODO: Add an int to keep track of how many neighbouring empty
squares there are and functions to increment/decrement this value.

Queens TODO: Add a float to keep track of Alpha scores and a function to
compute the Alpha score.