Queen Alpha Scores
Queens now store and calculate their local alpha scores (a measure of
mobility that the queen has - 0 means the queen is isolated from the
opponent)

Blank: renamed getPartitionedScore to getOmegaScore to be consistent
with the actual function variable.  Will add a comment in the code that
explains that this function is a measure of how close the game is to the
"filling" phase.

Gameboard:  added a method to calculate the alpha scores for all queens
on the board.  First resets the move distance counts for all blank
squares, recalculates them and uses the refreshed values to calculate
the alphas.  Also added a new test printing method.