AlphaBeta implemented
Minimax with alpha-beta pruning search implemented.  Right now does
about 6,500 nodes/second with a pruning rate of ~55%

To reach the 30 second cut off for a legal 2-ply search the algorithm
will have to be sped up to ~20,000 nodes/sec

A 2-ply search on a game 10 moves in takes about 4 seconds.  Will need
to dynamically increase the depth as the game progresses