Bug fix + Amazon game testing
Fixed a bug where numBlanks wasn't incrementing itself on an undoMove
(undo replaces an arrow with a blank, so the number of blanks needs to
go up by one) which was causing the queen and king distance algorithms
to break out of their calculations too early resulting in incorrect
values for the position.

Made the main class a playable demo for the game.  Be careful about
making illegal moves, there's no checking to make sure a move is well
formatted, only accepts moves in algebraic string format (a7, j10,
etc.), and will move the queen legally but won't fire an arrow illegally
(meaning you've only moved the queen but haven't made a shot this turn)

Type "stop" for the queen selection to stop the game and print the
encoded move history.